requirements.txt
asgiref==3.2.3
Django==2.1.15
django-rest-authtoken==1.2.4
django-rest-framework-auth==0.9.0
django-rest-registration==0.5.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
PyJWT==2.0.1
pytz==2019.3
sqlparse==0.3.0
typing==3.7.4.1
